refactor: 타입 수정으로 리팩토링
