+++ src/apis/hooks/myPots/useGetMyPotTodo.ts
@@ -1,30 +1,16 @@
-import { useQuery } from "@tanstack/react-query";
-import { GetTodos } from "apis/myPotAPI";
-import { GetTodoParams, Result } from "apis/types/todo";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { patchMyTodoStatus } from "apis/myPotAPI";
 
-const defaultResult: Result = {
-  totalPages: 0,
-  potName: "",
-  todos: [],
-  currentPage: 0,
-  totalElements: 0,
-};
+export const usePatchMyTodoStatus = () => {
+  const queryClient = useQueryClient();
 
-const useGetMyPotTodo = ({ potId, page, size }: GetTodoParams) => {
-  return useQuery({
-    queryKey: ["todos", potId, page, size],
-    queryFn: async () => {
-      const response = await GetTodos(potId, page, size);
-      return response.result ?? defaultResult;
+  return useMutation({
+    mutationFn: ({ potId, todoId }: { potId: number; todoId: number }) => patchMyTodoStatus(potId, todoId),
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["todos"] });
     },
-    select: (result) => ({
-      title: result.potName,
-      nickname: result.todos?.[0]?.userNickname ?? "",
-      taskCount: result.todos?.[0]?.todoCount ?? 0,
-      todos: result.todos,
-      totalElements: result.totalElements,
-    }),
+    onError: (error) => {
+      console.error("할 일 상태 변경 실패:", error);
+    },
   });
 };
-
-export default useGetMyPotTodo;
