+++ src/apis/hooks/myPots/usePatchMyTodo.ts
@@ -1,17 +1,17 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { patchTodo } from "apis/myPotAPI";
-import { TodoUpdateRequest } from "apis/types/todo";
+import { Todo } from "apis/types/myPot"; 
 
-export const usePatchMyTodo = () => {
+export const usePatchTodo = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: ({ potId, data }: TodoUpdateRequest) => patchTodo({ potId, data }), 
+    mutationFn: ({ potId, data }: { potId: number; data: Todo[] }) => patchTodo({ potId, data }),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ["todos"] });
     },
     onError: (error) => {
-      console.error("Error during patchTodo:", error);
+      console.error("할 일 업데이트 실패:", error);
     },
   });
 };
