+++ src/apis/myPotAPI.ts
@@ -1,13 +1,13 @@
 import { authApiGet, authApiPatch } from "./apiUtils";  
 import { ApiResponse } from "./types/response";  
-import { Result } from "./types/todo";
+import { Result, Todo } from "./types/myPot"; 
 
 export const GetTodos = async (potId: number, page: number = 1, size: number = 3): Promise<ApiResponse<Result>> => {
   const params = { page, size };
   return authApiGet<Result>(`/my-pots/${potId}/todos`, params);  
 };
 
-export const patchTodo = async ({ potId, data }: { potId: number; data: { todoId: number | null, content: string, status: string }[] }) 
+export const patchTodo = async ({ potId, data }: { potId: number; data: Todo[] }) 
   : Promise<ApiResponse<Result>> => {
   return authApiPatch<Result>(`/my-pots/${potId}/todos`, data);
 };
