--- src/apis/types/todo.ts
@@ -1,38 +1,0 @@
-export interface Todo {
-  todoId: number;
-  content: string;
-  status: string;
-}
-
-export interface User {
-  userNickname: string;
-  userRole: string;
-  userId: number;
-  todoCount: number;
-  todos: Todo[];
-}
-
-export interface Result {
-  totalPages: number;
-  potName: string;
-  todos: User[];
-  currentPage: number;
-  totalElements: number;
-}
-
-export interface GetTodoParams {
-  potId: number;
-  page: number;
-  size: number;
-}
-
-export interface TodoUpdateRequest {
-  potId: number;
-  data: TodoItem[];
-}
-
-export interface TodoItem {
-  todoId: number | null;
-  content: string;
-  status: string;
-}
